requirements.txt
streamlit
google.genarativeai
cryptography
os